requirements.txt
# Core dependencies
snowflake-snowpark-python>=1.11.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0.0
pandas>=2.0.0
cryptography>=41.0.0

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.8.0
isort>=5.13.0
autoflake>=2.2.0
pyupgrade>=3.15.0
pre-commit>=3.6.0

# Type stubs
types-PyYAML>=6.0.0
pandas-stubs>=2.0.0
